pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygadm"
version = "0.2.0"
description = "Easy access to administrative boundary defined by GADM from Python scripts"
keywords = ["Python"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.7.10"
dependencies = [
    "Deprecated",
    "geopandas",
    "pyarrow",
]

[[project.authors]]
name = "Pierrick Rambaud"
email = "pierrick.rambaud49@gmail.com"

[project.license]
text = "MIT"

[project.readme]
file = "README.rst"
content-type = "text/x-rst"

[project.urls]
Homepage = "https://github.com/12rambau/pygadm"
Download = "https://github.com/12rambau/pygadm/archive/${metadata:version}.tar.gz"

[project.optional-dependencies]
dev = [
    "pre-commit",
    "commitizen",
    "nox"
]
test = [
    "pytest",
    "pytest-sugar",
    "pytest-cov",
    "pytest-regressions"
]
doc = [
    "pydata-sphinx-theme",
    "sphinx-copybutton",
    "jupyter-sphinx",
    "sphinx-design",
    "ipyleaflet",
    "geemap",
    "sphinx-favicon",
]

[tool.setuptools]
include-package-data = true
packages = ["pygadm"]
license-files = ["LICENSE"]

[tool.setuptools.package-data]
pygadm = ["data/gadm_database.parquet", "data/gadm_continent.json"]

[tool.commitizen]
changelog_file = "CHANGELOG.md"
tag_format = "$major.$minor.$patch$prerelease"
update_changelog_on_bump = true
version = "0.2.0"
version_files = ["pyproject.toml:version", "pygadm/__init__.py:__version__"]

[tool.pytest.ini_options]
testpaths = "tests"

[tool.coverage.run]
source = ["pygadm"]

[tool.ruff]
fix = true
select = ["E", "F", "W", "I", "D", "RUF"]
ignore = [
    "E501",  # line too long | Black take care of it
    "W605",  # invalid escape sequence | we escape specific characters for sphinx
    "D212", # Multi-line docstring | we use a different convention, too late
]
exclude = ["setup.py"]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.pydocstyle]
convention = "google"

[tool.codespell]
skip = 'CHANGELOG.md,**/*.json'